# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for a test job in Avito",
        "title": "Avito-Rent API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/house/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new house. Requires moderator access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create House",
                "parameters": [
                    {
                        "description": "House details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HousePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "House created",
                        "schema": {
                            "$ref": "#/definitions/models.House"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/house/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve flats for a specific house. Requires authorization for both moderator and client.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get House Flats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flats retrieved",
                        "schema": {
                            "$ref": "#/definitions/utils.FlatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/house/{id}/subscribe": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Subscribe to updates for a specific house. Requires authorization for both moderator and client.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Subscribe to House",
                "parameters": [
                    {
                        "type": "string",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscription details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubscribePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscription successful",
                        "schema": {
                            "$ref": "#/definitions/utils.SubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Flat": {
            "description": "Represents a flat in the system",
            "type": "object",
            "properties": {
                "house_id": {
                    "description": "@Description Unique identifier for the house to which the flat belongs\n@Example 101",
                    "type": "integer"
                },
                "id": {
                    "description": "@Description Unique identifier for the flat\n@Example 1",
                    "type": "integer"
                },
                "price": {
                    "description": "@Description Price of the flat\n@Example 1200",
                    "type": "integer"
                },
                "rooms": {
                    "description": "@Description Number of rooms in the flat\n@Example 3",
                    "type": "integer"
                },
                "status": {
                    "description": "@Description Status of the flat\n@Example \"created\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FlatStatus"
                        }
                    ]
                }
            }
        },
...
        "models.FlatStatus": {
            "description": "Status of a flat",
            "type": "string",
            "enum": [
                "created",
                "approved",
                "declined",
                "on moderate"
            ],
            "x-enum-varnames": [
                "StatusCreated",
                "StatusApproved",
                "StatusDeclined",
                "StatusOnModeration"
            ]
        },
        "models.House": {
            "description": "House представляет собой структуру данных для хранения информации о доме.",
            "type": "object",
            "properties": {
                "address": {
                    "description": "@description Адрес дома\n@example \"123 Elm Street\"",
                    "type": "string"
                },
                "created_at": {
                    "description": "@description Дата создания записи\n@example \"2024-08-04T00:00:00Z\"",
                    "type": "string"
                },
                "developer": {
                    "description": "@description Разработчик или строитель дома\n@example \"XYZ Construction\"",
                    "type": "string"
                },
                "id": {
                    "description": "@description Идентификатор дома\n@example 1",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "@description Дата последнего обновления записи\n@example \"2024-08-04T00:00:00Z\"",
                    "type": "string"
                },
                "year": {
                    "description": "@description Год постройки\n@example 2020",
                    "type": "integer"
                }
            }
        },
        "models.HousePayload": {
            "description": "HousePayload представляет собой структуру данных для создания или обновления информации о доме.",
            "type": "object",
            "required": [
                "address",
                "year"
            ],
            "properties": {
                "address": {
                    "description": "@description Адрес дома\n@example \"123 Elm Street\"",
                    "type": "string"
                },
                "developer": {
                    "description": "@description Разработчик или строитель дома\n@example \"XYZ Construction\"",
                    "type": "string"
                },
                "year": {
                    "description": "@description Год постройки\n@example 2020",
                    "type": "integer"
                }
            }
        },
...
        "models.SubscribePayload": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "code": {
                    "description": "@example 400",
                    "type": "integer"
                },
                "message": {
                    "description": "@example \"Invalid input\""
                },
                "request_id": {
                    "description": "@example \"12345\""
                }
            }
        },
        "utils.FlatsResponse": {
            "description": "Response model for retrieving flats in a house",
            "type": "object",
            "properties": {
                "flats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Flat"
                    }
                }
            }
        },
...
        "utils.SubscriptionResponse": {
            "description": "Response model for subscription confirmation",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {}
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}